requirements.txt
pytest
configlookup
azure-storage-queue

